b4812

+++ client/cl_pred.c
@@ -53,9 +53,9 @@ void CL_CheckPredictionError (void)
 		VectorCopy (cl.frame.playerstate.pmove.origin, cl.predicted_origins[frame]);
 
 		// save for error itnerpolation
-		cl.prediction_error[0] = delta[0]*0.125;
-		cl.prediction_error[1] = delta[1]*0.125;
-		cl.prediction_error[2] = delta[2]*0.125;
+		cl.prediction_error[0] = delta[0]*0.125f;
+		cl.prediction_error[1] = delta[1]*0.125f;
+		cl.prediction_error[2] = delta[2]*0.125f;
 	}
 }
 
@@ -135,10 +135,10 @@ void CL_ClipMoveToEntities ( vec3_t start, vec3_t mins, vec3_t maxs, vec3_t end,
 			zd = 8*((ent->solid>>5) & 31);
 			zu = 8*((ent->solid>>10) & 63) - 32;
 
-			bmins[0] = bmins[1] = -x;
-			bmaxs[0] = bmaxs[1] = x;
-			bmins[2] = -zd;
-			bmaxs[2] = zu;
+			bmins[0] = bmins[1] = -(float)x;
+			bmaxs[0] = bmaxs[1] = (float)x;
+			bmins[2] = -(float)zd;
+			bmaxs[2] = (float)zu;
 
 			headnode = CM_HeadnodeForBox (bmins, bmaxs);
 			angles = vec3_origin;	// boxes don't rotate
@@ -330,13 +330,13 @@ void CL_PredictMovement (void)
 		{
 			case 3:
 				//get immediate results of this prediction vs last one
-				step = pm.s.origin[2] - cl.predicted_origin[2] * 8;
+				step = pm.s.origin[2] - (int)(cl.predicted_origin[2] * 8);
 
 				//r1ch: treat only some units as steps
 				if (((step > 62 && step < 66) || (step > 94 && step < 98) || (step > 126 && step < 130)) && !VectorCompare (pm.s.velocity, vec3_origin) && (pm.s.pm_flags & PMF_ON_GROUND))
 				{
-					cl.predicted_step = step * 0.125;
-					cl.predicted_step_time = cls.realtime - cls.frametime * 500;
+					cl.predicted_step = step * 0.125f;
+					cl.predicted_step_time = cls.realtime - (int)(cls.frametime * 500);
 				}
 				break;
 
@@ -351,8 +351,8 @@ void CL_PredictMovement (void)
 
 				if (last_step_frame != current && step > 63 && step < 160 && (pm.s.pm_flags & PMF_ON_GROUND) )
 				{
-					cl.predicted_step = step * 0.125;
-					cl.predicted_step_time = cls.realtime - cls.frametime * 500;
+					cl.predicted_step = step * 0.125f;
+					cl.predicted_step_time = cls.realtime - (int)(cls.frametime * 500);
 					last_step_frame = current;
 				}
 				break;
@@ -380,15 +380,15 @@ void CL_PredictMovement (void)
 		step = pm.s.origin[2] - oldz;
 		if (step > 63 && step < 160 && (pm.s.pm_flags & PMF_ON_GROUND) )
 		{
-			cl.predicted_step = step * 0.125;
-			cl.predicted_step_time = cls.realtime - cls.frametime * 500;
+			cl.predicted_step = step * 0.125f;
+			cl.predicted_step_time = cls.realtime - (int)(cls.frametime * 500);
 		}
 	}
 
 	// copy results out for rendering
-	cl.predicted_origin[0] = pm.s.origin[0]*0.125;
-	cl.predicted_origin[1] = pm.s.origin[1]*0.125;
-	cl.predicted_origin[2] = pm.s.origin[2]*0.125;
+	cl.predicted_origin[0] = pm.s.origin[0]*0.125f;
+	cl.predicted_origin[1] = pm.s.origin[1]*0.125f;
+	cl.predicted_origin[2] = pm.s.origin[2]*0.125f;
 
 	VectorCopy (pm.viewangles, cl.predicted_angles);
 }
